Refactor order and views
Add repository for orders and made authorization for it. Remove views blocks into separate files. Moved price into own trait

+++ database/migrations/2021_02_20_113349_create_order_carts_table.php
@@ -19,7 +19,7 @@ class CreateOrderCartsTable extends Migration
             $table->unsignedBigInteger('product_id');
             $table->integer('quantity');
             // keep it for history, since the price of product might change
-            $table->integer('price');
+            $table->unsignedBigInteger('price');
             $table->timestamps();
 
             $table->foreign('order_id')
